Remove old ASTVisualizer
